requirements.txt
imapclient
pyzmail36
pandas
openpyxl
python-dotenv
tzdata
msal
requests
